WIP endpoint tests
